+++ R/Connection.R
@@ -1,4 +1,5 @@
 #' @include Driver.R
+#' @include dplyr_integration.R
 NULL
 
 #' Athena Connection Methods
@@ -339,21 +340,58 @@ setMethod("dbDataType", c("AthenaConnection", "data.frame"), function(dbObj, obj
 #' @docType methods
 NULL
 
+# import DBI quote_string method
+dbi_quote = methods::getMethod("dbQuoteString", c("DBIConnection", "character"), asNamespace("DBI"))
+
+detect_date <- function(x, try_format = c("%Y-%m-%d", "%Y/%m/%d")){
+  return(all_dates = all(try(as.Date(x, tryFormats = try_format), silent=T) == x) & all(nchar(x) == 10))
+}
+
+detect_date_time <- function(x){
+  timestamp_fmt = c("%Y-%m-%d %H:%M:%OS", "%Y/%m/%d %H:%M:%OS", "%Y-%m-%d %H:%M", "%Y/%m/%d %H:%M")
+  return(all(try(as.POSIXct(x, tryFormats = timestamp_fmt), silent=T) == x))
+}
+
 #' @rdname dbQuote
 #' @export
 setMethod(
   "dbQuoteString", c("AthenaConnection", "character"),
   function(conn, x, ...) {
-    # Optional
-    getMethod("dbQuoteString", c("DBIConnection", "character"), asNamespace("DBI"))(conn, x, ...)
-  })
+    if(identical(dbplyr_env$major, 2L)){
+      all_ts <- detect_date_time(x)
+      all_dates <- detect_date(x)
+      if(all_dates & !is.na(all_dates)) {
+        return(paste0('date ', dbi_quote(conn, strftime(x, "%Y-%m-%d"), ...)))
+      } else if (all_ts & !is.na(all_ts)){
+        return(paste0('timestamp ', dbi_quote(conn, strftime(x, "%Y-%m-%d %H:%M:%OS3"), ...)))
+      }
+    }
+    return(dbi_quote(conn, x, ...))
+})
+
+#' @rdname dbQuote
+#' @export
+setMethod(
+  "dbQuoteString", c("AthenaConnection", "POSIXct"),
+  function(conn, x, ...) {
+    x <- strftime(x, "%Y-%m-%d %H:%M:%OS3")
+    paste0('timestamp ', dbi_quote(conn, x, ...))
+})
+
+#' @rdname dbQuote
+#' @export
+setMethod(
+  "dbQuoteString", c("AthenaConnection", "Date"),
+  function(conn, x, ...) {
+    paste0('date ', dbi_quote(conn, strftime(x, "%Y-%m-%d"), ...))
+})
 
 #' @rdname dbQuote
 #' @export
 setMethod(
   "dbQuoteIdentifier", c("AthenaConnection", "SQL"),
-  getMethod("dbQuoteIdentifier", c("DBIConnection", "SQL"), asNamespace("DBI")))
-
+  getMethod("dbQuoteIdentifier", c("DBIConnection", "SQL"), asNamespace("DBI"))
+)
 
 #' List Athena Tables
 #'
@@ -695,10 +733,26 @@ setMethod(
            ...){
     con_error_msg(conn, msg = "Connection already closed.")
     stopifnot(is.logical(statistics), is.logical(unload))
-    rs <- dbSendQuery(conn, statement = statement, unload = unload)
-    if(statistics) print(dbStatistics(rs))
-    out <- dbFetch(res = rs, n = -1, ...)
-    dbClearResult(rs)
+    
+    # dbplyr v2 support: dbplyr class ident
+    if(!inherits(statement, "ident")) {
+      rs <- dbSendQuery(conn, statement = statement, unload = unload)
+      if(statistics) print(dbStatistics(rs))
+      out <- dbFetch(res = rs, n = -1, ...)
+      dbClearResult(rs)
+    } else {
+      # Create an empty table using AWS GLUE to retrieve column names
+      field_names <- athena_query_fields_ident(conn, statement)
+      empty_shell <- rep(list(character()), length(field_names))
+      names(empty_shell) <- field_names
+      
+      if(inherits(athena_option_env[["file_parser"]], "athena_data.table")){
+        out <- as.data.table(empty_shell)
+      } else {
+        as_tibble <- pkg_method("as_tibble", "tibble")
+        out <- as_tibble(empty_shell)
+      }
+    }
     return(out)
 })