+++ R/zzz.R
@@ -1,13 +1,23 @@
 .onLoad <- function(libname, pkgname) {
   readr_check()
-  register_s3_method("dplyr", "db_desc", "AthenaConnection")
+  dbplyr_version()
+  if(identical(dbplyr_env$major, 1L)){
+    register_s3_method("dplyr", "db_desc", "AthenaConnection")
+    register_s3_method("dbplyr", "db_explain", "AthenaConnection")
+    register_s3_method("dbplyr", "db_query_fields", "AthenaConnection")
+    register_s3_method("dbplyr", "sql_translate_env", "AthenaConnection")
+    register_s3_method("dbplyr", "sql_escape_string", "AthenaConnection")
+  } else if(identical(dbplyr_env$major, 2L)){
+    register_s3_method("dbplyr", "dbplyr_edition", "AthenaConnection")
+    register_s3_method("dbplyr", "db_connection_describe", "AthenaConnection")
+    register_s3_method("dbplyr", "sql_query_explain", "AthenaConnection")
+    register_s3_method("dbplyr", "sql_query_fields", "AthenaConnection")
+    register_s3_method("dbplyr", "sql_translation", "AthenaConnection")
+    register_s3_method("dbplyr", "sql_escape_date", "AthenaConnection")
+    register_s3_method("dbplyr", "sql_escape_datetime", "AthenaConnection")
+  }
   register_s3_method("dbplyr", "db_compute", "AthenaConnection")
-  register_s3_method("dplyr", "db_save_query", "AthenaConnection")
   register_s3_method("dbplyr", "db_copy_to", "AthenaConnection")
-  register_s3_method("dbplyr", "sql_translate_env", "AthenaConnection")
-  register_s3_method("dbplyr", "sql_escape_string", "AthenaConnection")
-  register_s3_method("dbplyr", "db_explain", "AthenaConnection")
-  register_s3_method("dbplyr", "db_query_fields", "AthenaConnection")
 }
 
 register_s3_method <- function(pkg, generic, class, fun = NULL) {
@@ -38,3 +48,17 @@ readr_check <- function() {
   if (!nzchar(system.file(package = "readr")))
     packageStartupMessage("Info: For extra speed please install `readr`.")
 }
+
+dbplyr_version <- function(){
+  if (nzchar(system.file(package = "dbplyr"))){
+    dbplyr_v = packageVersion("dbplyr")
+    dbplyr_env$major <- dbplyr_v$major
+    dbplyr_env$minor <- dbplyr_v$minor
+  } else {
+    # default to minimum supported dbplyr version
+    dbplyr_env$major = 1L
+    dbplyr_env$minor = 4L
+  }
+}
+
+dbplyr_env <- new.env(parent=emptyenv())